+++ src/components/main/TopNav.tsx
@@ -1,8 +1,8 @@
-import { handleLogout } from '@/api/TokenApi';
 import { Link } from 'react-router';
 import { Button } from '../ui/button';
 import { useUserTokenTemp } from '@/hooks/useUserTokenTemp';
 import { loginApiTemp } from '@/api/loginApiTemp';
+import { handleLogout } from '@/pages/user/handleLogout';
 
 export default function TopNav() {
   const accessToken = localStorage.getItem('accessToken');
